# Elevance/specs/001-prior-auth-assistant/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Prior Authorization Evidence Assistant API
  version: 1.0.0
  description: API Gateway orchestrating the 5-agent pipeline for PA cases.
paths:
  /cases:
    post:
      summary: Create a new PA case
      operationId: createCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [member_id, provider_id, request_type]
              properties:
                member_id:
                  type: string
                provider_id:
                  type: string
                request_type:
                  type: string
                cpt_hcpcs_codes:
                  type: array
                  items:
                    type: string
                icd_10_codes:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Case created
          content:
            application/json:
              schema:
                type: object
                properties:
                  case_id:
                    type: string
                  status:
                    type: string
                  next_step:
                    type: string
  /cases/{case_id}/documents:
    post:
      summary: Upload attachments (PDFs, images) for a case
      operationId: uploadDocuments
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Documents accepted for async OCR and indexing
...